Add support for `accessibleM` and `accessibleMM`
