+++ src/main/resources/META-INF/plugin.xml
@@ -28,6 +28,8 @@
 
     <extensions defaultExtensionNs="org.intellij.scala">
         <syntheticMemberInjector implementation="zio.intellij.synthetic.macros.ModulePatternAccessible"/>
+        <syntheticMemberInjector implementation="zio.intellij.synthetic.macros.ModulePatternAccessibleM"/>
+        <syntheticMemberInjector implementation="zio.intellij.synthetic.macros.ModulePatternAccessibleMM"/>
         <syntheticMemberInjector implementation="zio.intellij.synthetic.macros.MockableInjector"/>
     </extensions>
