+++ src/main/scala/zio/intellij/synthetic/macros/ModulePatternAccessibleMM.scala
@@ -1,0 +1,6 @@
+package zio.intellij.synthetic.macros
+
+class ModulePatternAccessibleMM extends ModulePatternAccessibleMBase {
+  override protected val macroName: String        = "zio.macros.accessibleMM"
+  override protected val expectedTypeParams: Long = 2
+}
